Add a wildcard handler to redirect unhandled welsh urls
